First branch commit

+++ routes/web.php
@@ -29,26 +29,27 @@ Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
+
+    //Rutas del apartado Cocina
+    Route::get('/cocina', [CocinaController::class , 'index'])->name('cocina')->middleware('cocinero');
+    Route::post('/cocina/{id}', [CocinaController::class , '@listo'])->name('cocina-listo')->middleware('cocinero');
+
+    //Rutas del apartado camarero para crear comanda, servirlas y marcar pedidos como pagado.
+    Route::get('/camarero', [CamareroController::class , 'index'])->name('camarero')->middleware('camarero');
+    Route::post('/camarero', [CamareroController::class , 'newComanda'])->name('camarero')->middleware('camarero');
+    Route::post('/camarero/{id}', [CamareroController::class , 'servido'])->name('camarero-servido')->middleware('camarero');
+    Route::post('/pagado/{id}', [CamareroController::class , 'pagado'])->name('camarero-pagado')->middleware('camarero');
+
+
+    //Rutas del apartado administracion para poder crear, editar, y eliminar platos.
+    Route::get('/admin', [AdminController::class , 'index'])->name('admin')->middleware('admin');
+    Route::post('/admin', [AdminController::class , 'addPlato'])->middleware('admin');
+    Route::get('/plato/{id}', [AdminController::class , 'showPlato'])->name('plato-edit')->middleware('admin');
+    Route::patch('/plato/{id}', [AdminController::class , 'updatePlato'])->name('plato-update')->middleware('admin');
+    Route::delete('/plato/{id}', [AdminController::class , 'borrarPlato'])->name('plato-destroy')->middleware('admin');
 });
 
-//Rutas del apartado Cocina
-Route::get('/cocina', CocinaController::class.'@index')->name('cocina');
-Route::post('/cocina/{id}', CocinaController::class.'@listo')->name('cocina-listo');
-
-//Rutas del apartado camarero para crear comanda, servirlas y marcar pedidos como pagado.
-Route::get('/camarero', CamareroController::class.'@index')->name('camarero');
-Route::post('/camarero', CamareroController::class.'@newComanda')->name('camarero');
-Route::post('/camarero/{id}', CamareroController::class.'@servido')->name('camarero-servido');
-Route::post('/pagado/{id}', CamareroController::class.'@pagado')->name('camarero-pagado');
 
 
-//Rutas del apartado administracion para poder crear, editar, y eliminar platos.
-Route::get('/admin', AdminController::class.'@index')->name('admin');
-Route::post('/admin', AdminController::class.'@addPlato');
-Route::get('/plato/{id}',AdminController::class.'@showPlato')->name('plato-edit');
-Route::patch('/plato/{id}',AdminController::class.'@updatePlato')->name('plato-update');
-Route::delete('/plato/{id}',AdminController::class.'@borrarPlato')->name('plato-destroy');
 
-
-require __DIR__.'/auth.php';
-
+require __DIR__ . '/auth.php';
